Reset `ElementConfigurationContainer` and `ElementConfigurationSection` changes which got lost during previous rebase.
